refactor: align utils split tests and shared truncate types
