+++ src/architecture.test.ts
@@ -7,7 +7,7 @@ describe("architecture boundaries", () => {
     const srcDir = join(import.meta.dir);
     const metadata = readFileSync(join(srcDir, "metadata.ts"), "utf8");
 
-    expect(metadata).not.toContain('from "./utils"');
+    expect(metadata).not.toMatch(/\bfrom\s+["']\.\/utils["']/);
     expect(existsSync(join(srcDir, "utils.ts"))).toBe(false);
   });
 });
