+++ src/content/types.ts
@@ -1,0 +1,1 @@
+export type TruncateMethod = "head_only" | "head_tail" | "heading";
